First draft of new backlight timeout
I'm not sure if I caused more issues with this then I fixed. The screen
and buttons don't respond exactly right, and the backlight timeout
doesn't work properly in o-scope mode at all. I'm starting to think the
RTC interrupt is used for more than the 1 second ticks. And here I
thought this would be easy!